Added class OBSObject.
This commit adds the OBSObject class, to be inherited by all Library
classes.
An OBSObject has all proprieties and methods that are common to all
Library objects.
For now, it contains a reference to its client, a method to initiate an
object with a client, the default init method is overridden to return
nil, and it contains a tag and an identifier to be used by the client.